Move some classes/packages around
