+++ app/src/main/java/com/paulmandal/atak/forwarder/comm/commhardware/meshtastic/MeshtasticDeviceSwitcher.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.forwarder.comm.commhardware;
+package com.paulmandal.atak.forwarder.comm.commhardware.meshtastic;
 
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
@@ -12,7 +12,6 @@ import android.util.Log;
 
 import com.geeksville.mesh.IMeshService;
 import com.paulmandal.atak.forwarder.Config;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDevice;
 
 import java.util.ArrayList;
 import java.util.HashMap;
