--- app/src/main/java/com/paulmandal/atak/forwarder/persistence/ConfigState.java
@@ -1,33 +1,0 @@
-package com.paulmandal.atak.forwarder.persistence;
-
-//public class ConfigState {
-//    public MeshtasticDevice commDevice;
-//    public int pliMaxFrequencyS;
-//    public int duplicateMessagesTtlM;
-//    public int pliHopLimit;
-//    public int chatHopLimit;
-//    public int otherHopLimit;
-//    public String channelName;
-//    public int channelMode;
-//    public byte[] channelPsk;
-//    public int trackerPliIntervalS;
-//    public int trackerScreenOffTimeS;
-//    public int trackerTeamIndex;
-//    public int trackerRoleIndex;
-//
-//    /**
-//     * These should match the values in preferences.xml
-//     */
-//    public final int defaultPliMaxFrequencyS      = 30;
-//    public final int defaultDuplicateMessagesTtlM = 20;
-//    public final int defaultPliHopLimit           = 3;
-//    public final int defaultChatHopLimit          = 3;
-//    public final int defaultOtherHopLimit         = 3;
-//    public final String defaultChannelName        = "Default";
-//    public final int defaultChannelMode           = 0;
-//    public final String defaultChannelPskStr      = "";
-//    public final int defaultTrackerPliIntervalS   = 20;
-//    public final int defaultTrackerScreenOffTimeS = 5;
-//    public final int defaultTrackerTeamIndex      = 0;
-//    public final int defaultTrackerRoleIndex      = 0;
-//}
