--- app/src/main/java/com/paulmandal/atak/forwarder/persistence/PreferencesReader.java
@@ -1,9 +1,0 @@
-package com.paulmandal.atak.forwarder.persistence;
-
-public class PreferencesReader {
-//    public static ConfigState readConfigFromPreferences(SharedPreferences preferences) {
-//        ConfigState configState = new ConfigState();
-//        // TODO: Read the whole thing w/ defaults?
-//        return configState;
-//    }
-}
